+++ lesson_03/lesson_3_task1.py
@@ -1,15 +1,13 @@
-class User:
-    def __init__(self, first_name, last_name):
-        self.first_name = first_name
-        self.last_name = last_name
-        self.fullName = f"{self.first_name} {self.last_name}"
+from user_m import User
 
-    def print_first_name(self):
-        print(f'Меня зовут: {self.first_name}')
+my_user = User("Aglaiya", "Pervenceva")
 
-    def print_Last_name(self):
-        print(f'Моя фамилия: {self.last_name}')
+my_user.print_first_name()
+my_user.print_Last_name()
+my_user.print_fullName()
 
 
-    def print_fullName(self):
-        print(f'Мои имя и фамилия: {self.fullName}')
+
+
+
+
