+++ lesson_03/user_m.py
@@ -1,7 +1,15 @@
-from lesson_3_task1 import User
+class User:
+    def __init__(self, first_name, last_name):
+        self.first_name = first_name
+        self.last_name = last_name
+        self.fullName = f"{self.first_name} {self.last_name}"
 
-my_user = User("Aglaiya", "Pervenceva")
+    def print_first_name(self):
+        print(f'Меня зовут: {self.first_name}')
 
-my_user.print_first_name()
-my_user.print_Last_name()
-my_user.print_fullName()
+    def print_Last_name(self):
+        print(f'Моя фамилия: {self.last_name}')
+
+
+    def print_fullName(self):
+        print(f'Мои имя и фамилия: {self.fullName}')
